# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de proveedores",
        "description": "Documentación de la API para gestión de proveedores y Productoos.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8080",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/Productos": {
            "get": {
                "tags": [
                    "Productoos"
                ],
                "summary": "Listar Productoos",
                "operationId": "ff0b4a6165eba88343b119b397f54468",
                "responses": {
                    "200": {
                        "description": "Lista de Productoos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Producto"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Productoos"
                ],
                "summary": "Crear Productoo",
                "operationId": "934969049f06975f7e91ea42a28a086b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Producto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Productoo creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Productos/{id}": {
            "get": {
                "tags": [
                    "Productoos"
                ],
                "summary": "Obtener Productoo por ID",
                "operationId": "3bcf4435395d7e4f59684622ce37c95e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Productoo encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Productoos"
                ],
                "summary": "Actualizar Productoo",
                "operationId": "b646a2fb0e838123526151f272c1108c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Producto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Productoo actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Productoos"
                ],
                "summary": "Eliminar Productoo",
                "operationId": "bc22f4873704423dfea7f572d0663f3a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Productoo eliminado"
                    }
                }
            }
        },
        "/api/proveedores": {
            "get": {
                "tags": [
                    "proveedores"
                ],
                "summary": "Listar proveedores",
                "operationId": "7f6ed20dc86f1682aa7e31d71afa37df",
                "responses": {
                    "200": {
                        "description": "Lista de proveedores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Proveedor"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "proveedores"
                ],
                "summary": "Crear proveedor",
                "operationId": "a1046f3d7a3c804651bf420761995e57",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Proveedor"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proveedor creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proveedor"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/proveedores/{id}": {
            "get": {
                "tags": [
                    "proveedores"
                ],
                "summary": "Obtener proveedor por ID",
                "operationId": "5aa647e1f6ead15a2c088ecc1fbb12bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proveedor encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proveedor"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "proveedores"
                ],
                "summary": "Actualizar proveedor",
                "operationId": "28ff17fc2a31c61745cbe9a18da5aa32",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Proveedor"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proveedor actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proveedor"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "proveedores"
                ],
                "summary": "Eliminar proveedor",
                "operationId": "2b9dc197602ee96f15138b4ad2024fab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Proveedor eliminado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Catalogo": {
                "required": [
                    "nombre",
                    "proveedor_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Catálogo 2024"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Catálogo anual de Productoos",
                        "nullable": true
                    },
                    "proveedor_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "photo_url": {
                        "type": "string",
                        "example": "https://dominio.com/storage/catalogos/2024.jpg",
                        "nullable": true
                    },
                    "Proveedor": {
                        "$ref": "#/components/schemas/Proveedor"
                    }
                },
                "type": "object"
            },
            "Categoria": {
                "required": [
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Herramientas eléctricas"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Categoría de herramientas eléctricas",
                        "nullable": true
                    },
                    "photo_url": {
                        "type": "string",
                        "example": "https://dominio.com/storage/categoria/electricas.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Linea": {
                "required": [
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Industrial"
                    },
                    "photo_url": {
                        "type": "string",
                        "example": "https://dominio.com/storage/lineas/industrial.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Marca": {
                "required": [
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Bosch"
                    },
                    "photo_url": {
                        "type": "string",
                        "example": "https://dominio.com/storage/marcas/bosch.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Producto": {
                "required": [
                    "catalogo_id",
                    "nombre",
                    "sku"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "catalogo_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Taladro Bosch"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Herramienta eléctrica",
                        "nullable": true
                    },
                    "sku": {
                        "type": "string",
                        "example": "SKU-001"
                    },
                    "precio": {
                        "type": "number",
                        "format": "float",
                        "example": 199.99
                    },
                    "cantidad_disponible": {
                        "type": "integer",
                        "example": 10
                    },
                    "marca_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "linea_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    },
                    "photo_url": {
                        "type": "string",
                        "example": "https://dominio.com/storage/Productos/sku-001.jpg",
                        "nullable": true
                    },
                    "Catalogo": {
                        "$ref": "#/components/schemas/Catalogo"
                    },
                    "Marca": {
                        "$ref": "#/components/schemas/Marca"
                    },
                    "Linea": {
                        "$ref": "#/components/schemas/Linea"
                    },
                    "categoria": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Categoria"
                        }
                    }
                },
                "type": "object"
            },
            "Proveedor": {
                "required": [
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Suministros Industriales S.A."
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Av. Central 123",
                        "nullable": true
                    },
                    "telefono": {
                        "type": "string",
                        "example": "555-1234",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "example": "contacto@proveedor.com",
                        "nullable": true
                    },
                    "photo_url": {
                        "type": "string",
                        "example": "https://dominio.com/storage/proveedores/si.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}